Move githubGrant.js to src/Auth0 (Model, not View)
src/Components is View in the loose MVC layout; the grant record is
Auth0-login-flow state, so it belongs beside the other Auth0 plumbing.
Also inventory its storage keys in src/connections/README.md and note the
module dies with the legacy federated-token path.
